Removed the '_lib.py' and included the functions directly
